Moved mock data out of __tests__ folder
